rename .js files to .jsx for consistency with React guidelines
