+++ js/app.jsx
@@ -1,4 +1,4 @@
-import Hello from './Hello'
+import Hello from './Hello.jsx'
 import React from 'react'
 import ReactDOM from 'react-dom'
